Moves AddressForm to a separate folder
