+++ packages/client/src/comp/Address/Form.tsx
@@ -1,19 +1,19 @@
 import React from 'react'
 import Input from 'muicss/lib/react/input'
 
-import { RoundedButton } from './util/Button'
-import { client } from '../lib/trpc'
-import { AddressContainer, ContactContainer, FetchingDataContainer } from '../lib/unstated'
-import { useControlRef } from './util/ControlRef'
+import { RoundedButton } from '../util/Button'
+import { client } from '../../lib/trpc'
+import { AddressContainer, ContactContainer, FetchingDataContainer } from '../../lib/unstated'
+import { useControlRef } from '../util/ControlRef'
 import { TimeoutError } from '@tianhuil/simple-trpc/dist/timedFetch'
-import { BaseInput } from './util/Input'
-import { StatusReport } from './status/StatusReport'
+import { BaseInput } from '../util/Input'
+import { StatusReport } from '../status/StatusReport'
 import { useParams } from 'react-router-dom'
-import { useAppHistory } from '../lib/path'
+import { useAppHistory } from '../../lib/path'
 import styled from 'styled-components'
-import { sampleAddresses, ImplementedState, getState } from '../common'
-import { AppForm } from './util/Form'
-import { Unidentified } from './status/Status'
+import { sampleAddresses, ImplementedState, getState } from '../../common'
+import { AppForm } from '../util/Form'
+import { Unidentified } from '../status/Status'
 import { toast } from 'react-toastify'
 
 const FlexBox = styled.div`
